% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-series-circles.R
\docType{data}
\name{GeomSeriesCircles}
\alias{GeomSeriesCircles}
\title{See \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}}
\description{
See \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}
}
\keyword{datasets}
