% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtricks-package.R
\docType{package}
\name{ggtricks-package}
\alias{ggtricks}
\alias{ggtricks-package}
\title{ggtricks: Create Sector and Other Charts Easily Using Grammar of Graphics}
\description{
A collection of several geom to create graphics, using "ggplot2" and the Cartesian coordinate system. You use the familiar mapping 'Grammar of Graphics' without the need to do another transformation into polar cordinates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AbdoulMa/ggtricks}
  \item Report bugs at \url{https://github.com/AbdoulMa/ggtricks/issues}
}

}
\author{
\strong{Maintainer}: Abdoul ISSA BIDA \email{contact@abdoulblog.com}

}
\keyword{internal}
