% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-donut-slice.R
\docType{data}
\name{StatDonutSlice}
\alias{StatDonutSlice}
\title{See \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}}
\description{
See \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}
}
\keyword{datasets}
