% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-series-circles.R
\docType{data}
\name{StatSeriesCircles}
\alias{StatSeriesCircles}
\title{See \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}}
\description{
See \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}
}
\keyword{datasets}
